Refactor and fix wrong logIndex
